Refactored main.js
